game-logic-testing-zone-of-Ann cleaned up so only Board.jsx remains
