--- Game-Logic-Testing-Zone-of-Ann/Additional-Rules.jsx
@@ -1,12 +1,0 @@
-// Thinking Space
-
-// Player can only click on one piece on their turn
-    // Click to select piece
-        // Player can only select their pieces
-    // Click to deselect piece ???
-    // Click to select destination for selected piece
-        // Check if move is valid, etc.
-        // Piece stays selected
-        // Loop until no valid moves
-    // Double click to end turn
-        // Deselect piece
