+++ Game-Logic-Testing-Zone-of-Ann/Board.jsx
@@ -1,4 +1,4 @@
-// Thinking Space
+// Testing Space
 
 import { useState } from 'react';
 
@@ -40,6 +40,8 @@ export default function Board() {
   const statusMessage = ["Black's Turn", "Red's Turn", "Black Wins!", "Red Wins!"]
   const [status, setStatus] = useState(statusMessage[0]);
 
+  const [ghostPosition, setGhostPosition] = useState();
+
   function handleClick(i) {
     const nextSquares = squares.slice();
 
@@ -65,10 +67,14 @@ export default function Board() {
       }
       if ((redCount === 0) && blackIsNext) {
         setStatus(statusMessage[2]);
-        console.log('Black wins!');
+        // console.log('Black wins!');
+        // TODO Add 1 to Black Player's Win Stat
+        // TODO Add 1 to Red Player's Loss Stat
       } else if ((blackCount === 0) && !blackIsNext) {
         setStatus(statusMessage[3]);
-        console.log('Red wins!');
+        // console.log('Red wins!');
+        // TODO Add 1 to Red Player's Win Stat
+        // TODO Add 1 to Black Player's Loss Stat
       }
     }
 
@@ -79,70 +85,2456 @@ export default function Board() {
         }
       }
     }
+
+    function getGhostPosition() {
+      for (let index = 0; index < nextSquares.length; index++) {
+        if ((nextSquares[index] === 'B') || (nextSquares[index] === 'R') || (nextSquares[index] === 'BK') || (nextSquares[index] === 'RK')) {
+          setGhostPosition(index);
+        }
+      }
+    }
+
+    function movementRules() {
+      // Single Black Piece Move Set
+      if ((ghostPosition === 1) && blackIsNext && (selectedPiece === '⚫')) {
+        // Ex. "B1" to ("A2" || ("C2" || "D3")
+        if ((nextSquares[8] === null) && (i === 8)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[10] === null) && (i === 10)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[19] === null) && ((nextSquares[10] === '🔴') || (nextSquares[10] === '❤️')) && (i === 19)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 1) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 3) && blackIsNext && (selectedPiece === '⚫')) {
+        if ((nextSquares[10] === null) && (i === 10)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[12] === null) && (i === 12)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[17] === null) && ((nextSquares[10] === '🔴') || (nextSquares[10] === '❤️')) && (i === 17)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[21] === null) && ((nextSquares[12] === '🔴') || (nextSquares[12] === '❤️')) && (i === 21)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 3) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 5) && blackIsNext && (selectedPiece === '⚫')) {
+        if ((nextSquares[12] === null) && (i === 12)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[14] === null) && (i === 14)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[19] === null) && ((nextSquares[12] === '🔴') || (nextSquares[12] === '❤️')) && (i === 19)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[23] === null) && ((nextSquares[14] === '🔴') || (nextSquares[14] === '❤️')) && (i === 23)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 5) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 7) && blackIsNext && (selectedPiece === '⚫')) {
+        if ((nextSquares[14] === null) && (i === 14)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[21] === null) && ((nextSquares[14] === '🔴') || (nextSquares[14] === '❤️')) && (i === 21)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 7) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 8) && blackIsNext && (selectedPiece === '⚫')) {
+        if ((nextSquares[17] === null) && (i === 17)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[26] === null) && ((nextSquares[17] === '🔴') || (nextSquares[17] === '❤️')) && (i === 26)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 8) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 10) && blackIsNext && (selectedPiece === '⚫')) {
+        if ((nextSquares[17] === null) && (i === 17)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[19] === null) && (i === 19)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[24] === null) && ((nextSquares[17] === '🔴') || (nextSquares[17] === '❤️')) && (i === 24)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[28] === null) && ((nextSquares[19] === '🔴') || (nextSquares[19] === '❤️')) && (i === 28)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 10) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 12) && blackIsNext && (selectedPiece === '⚫')) {
+        if ((nextSquares[19] === null) && (i === 19)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[21] === null) && (i === 21)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[26] === null) && ((nextSquares[19] === '🔴') || (nextSquares[19] === '❤️')) && (i === 26)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[30] === null) && ((nextSquares[21] === '🔴') || (nextSquares[21] === '❤️')) && (i === 30)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 12) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 14) && blackIsNext && (selectedPiece === '⚫')) {
+        if ((nextSquares[21] === null) && (i === 21)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[23] === null) && (i === 23)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[28] === null) && ((nextSquares[21] === '🔴') || (nextSquares[21] === '❤️')) && (i === 28)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 14) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 17) && blackIsNext && (selectedPiece === '⚫')) {
+        if ((nextSquares[24] === null) && (i === 24)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[26] === null) && (i === 26)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[35] === null) && ((nextSquares[26] === '🔴') || (nextSquares[26] === '❤️')) && (i === 35)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 17) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 19) && blackIsNext && (selectedPiece === '⚫')) {
+        if ((nextSquares[26] === null) && (i === 26)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[28] === null) && (i === 28)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[33] === null) && ((nextSquares[26] === '🔴') || (nextSquares[26] === '❤️')) && (i === 33)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[37] === null) && ((nextSquares[28] === '🔴') || (nextSquares[28] === '❤️')) && (i === 37)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 19) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 21) && blackIsNext && (selectedPiece === '⚫')) {
+        if ((nextSquares[28] === null) && (i === 28)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[30] === null) && (i === 30)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[35] === null) && ((nextSquares[28] === '🔴') || (nextSquares[28] === '❤️')) && (i === 35)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[39] === null) && ((nextSquares[30] === '🔴') || (nextSquares[30] === '❤️')) && (i === 39)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 21) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 23) && blackIsNext && (selectedPiece === '⚫')) {
+        if ((nextSquares[30] === null) && (i === 30)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[37] === null) && ((nextSquares[30] === '🔴') || (nextSquares[30] === '❤️')) && (i === 37)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 23) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 24) && blackIsNext && (selectedPiece === '⚫')) {
+        if ((nextSquares[33] === null) && (i === 33)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[42] === null) && ((nextSquares[33] === '🔴') || (nextSquares[33] === '❤️')) && (i === 42)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 24) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 26) && blackIsNext && (selectedPiece === '⚫')) {
+        if ((nextSquares[33] === null) && (i === 33)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[35] === null) && (i === 35)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[40] === null) && ((nextSquares[33] === '🔴') || (nextSquares[33] === '❤️')) && (i === 40)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[44] === null) && ((nextSquares[35] === '🔴') || (nextSquares[35] === '❤️')) && (i === 44)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 26) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 28) && blackIsNext && (selectedPiece === '⚫')) {
+        if ((nextSquares[35] === null) && (i === 35)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[37] === null) && (i === 37)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[42] === null) && ((nextSquares[35] === '🔴') || (nextSquares[35] === '❤️')) && (i === 42)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[46] === null) && ((nextSquares[37] === '🔴') || (nextSquares[37] === '❤️')) && (i === 46)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 28) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 30) && blackIsNext && (selectedPiece === '⚫')) {
+        if ((nextSquares[37] === null) && (i === 37)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[39] === null) && (i === 39)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[44] === null) && ((nextSquares[37] === '🔴') || (nextSquares[37] === '❤️')) && (i === 44)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 30) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 33) && blackIsNext && (selectedPiece === '⚫')) {
+        if ((nextSquares[40] === null) && (i === 40)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[42] === null) && (i === 42)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[51] === null) && ((nextSquares[42] === '🔴') || (nextSquares[42] === '❤️')) && (i === 51)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 33) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 35) && blackIsNext && (selectedPiece === '⚫')) {
+        if ((nextSquares[42] === null) && (i === 42)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[44] === null) && (i === 44)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[49] === null) && ((nextSquares[42] === '🔴') || (nextSquares[42] === '❤️')) && (i === 49)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[53] === null) && ((nextSquares[44] === '🔴') || (nextSquares[44] === '❤️')) && (i === 53)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 35) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 37) && blackIsNext && (selectedPiece === '⚫')) {
+        if ((nextSquares[44] === null) && (i === 44)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[46] === null) && (i === 46)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[51] === null) && ((nextSquares[44] === '🔴') || (nextSquares[44] === '❤️')) && (i === 51)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[55] === null) && ((nextSquares[46] === '🔴') || (nextSquares[46] === '❤️')) && (i === 55)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 37) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 39) && blackIsNext && (selectedPiece === '⚫')) {
+        if ((nextSquares[46] === null) && (i === 46)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[53] === null) && ((nextSquares[46] === '🔴') || (nextSquares[46] === '❤️')) && (i === 53)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 39) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 40) && blackIsNext && (selectedPiece === '⚫')) {
+        if ((nextSquares[49] === null) && (i === 49)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[58] === null) && ((nextSquares[49] === '🔴') || (nextSquares[49] === '❤️')) && (i === 58)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 40) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 42) && blackIsNext && (selectedPiece === '⚫')) {
+        if ((nextSquares[49] === null) && (i === 49)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[51] === null) && (i === 51)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[56] === null) && ((nextSquares[49] === '🔴') || (nextSquares[49] === '❤️')) && (i === 56)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[60] === null) && ((nextSquares[51] === '🔴') || (nextSquares[51] === '❤️')) && (i === 60)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 42) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 44) && blackIsNext && (selectedPiece === '⚫')) {
+        if ((nextSquares[51] === null) && (i === 51)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[53] === null) && (i === 53)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[58] === null) && ((nextSquares[51] === '🔴') || (nextSquares[51] === '❤️')) && (i === 58)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[62] === null) && ((nextSquares[53] === '🔴') || (nextSquares[53] === '❤️')) && (i === 62)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 44) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 46) && blackIsNext && (selectedPiece === '⚫')) {
+        if ((nextSquares[53] === null) && (i === 53)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[55] === null) && (i === 55)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[60] === null) && ((nextSquares[53] === '🔴') || (nextSquares[53] === '❤️')) && (i === 60)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 46) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 49) && blackIsNext && (selectedPiece === '⚫')) {
+        if ((nextSquares[56] === null) && (i === 56)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[58] === null) && (i === 58)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 49) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 51) && blackIsNext && (selectedPiece === '⚫')) {
+        if ((nextSquares[58] === null) && (i === 58)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[60] === null) && (i === 60)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 51) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 53) && blackIsNext && (selectedPiece === '⚫')) {
+        if ((nextSquares[60] === null) && (i === 60)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[62] === null) && (i === 62)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 53) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 55) && blackIsNext && (selectedPiece === '⚫')) {
+        if ((nextSquares[62] === null) && (i === 62)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 55) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      
+      // Kinged Black Piece Move Set
+      if ((ghostPosition === 1) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[8] === null) && (i === 8)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[10] === null) && (i === 10)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[19] === null) && ((nextSquares[10] === '🔴') || (nextSquares[10] === '❤️')) && (i === 19)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 1) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 3) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[10] === null) && (i === 10)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[12] === null) && (i === 12)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[17] === null) && ((nextSquares[10] === '🔴') || (nextSquares[10] === '❤️')) && (i === 17)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[21] === null) && ((nextSquares[12] === '🔴') || (nextSquares[12] === '❤️')) && (i === 21)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 3) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 5) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[12] === null) && (i === 12)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[14] === null) && (i === 14)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[19] === null) && ((nextSquares[12] === '🔴') || (nextSquares[12] === '❤️')) && (i === 19)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[23] === null) && ((nextSquares[14] === '🔴') || (nextSquares[14] === '❤️')) && (i === 23)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 5) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 7) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[14] === null) && (i === 14)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[21] === null) && ((nextSquares[14] === '🔴') || (nextSquares[14] === '❤️')) && (i === 21)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 7) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 8) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[1] === null) && (i === 1)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[17] === null) && (i === 17)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[26] === null) && ((nextSquares[17] === '🔴') || (nextSquares[17] === '❤️')) && (i === 26)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 8) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 10) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[1] === null) && (i === 1)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[3] === null) && (i === 3)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[17] === null) && (i === 17)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[19] === null) && (i === 19)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[24] === null) && ((nextSquares[17] === '🔴') || (nextSquares[17] === '❤️')) && (i === 24)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[28] === null) && ((nextSquares[19] === '🔴') || (nextSquares[19] === '❤️')) && (i === 28)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 10) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 12) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[3] === null) && (i === 3)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[5] === null) && (i === 5)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[19] === null) && (i === 19)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[21] === null) && (i === 21)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[26] === null) && ((nextSquares[19] === '🔴') || (nextSquares[19] === '❤️')) && (i === 26)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[30] === null) && ((nextSquares[21] === '🔴') || (nextSquares[21] === '❤️')) && (i === 30)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 12) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 14) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[5] === null) && (i === 5)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[7] === null) && (i === 7)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[21] === null) && (i === 21)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[23] === null) && (i === 23)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[28] === null) && ((nextSquares[21] === '🔴') || (nextSquares[21] === '❤️')) && (i === 28)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 14) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 17) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[8] === null) && (i === 8)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[10] === null) && (i === 10)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[3] === null) && ((nextSquares[10] === '🔴') || (nextSquares[10] === '❤️')) && (i === 3)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[24] === null) && (i === 24)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[26] === null) && (i === 26)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[35] === null) && ((nextSquares[26] === '🔴') || (nextSquares[26] === '❤️')) && (i === 35)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 17) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 19) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[10] === null) && (i === 10)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[12] === null) && (i === 12)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[1] === null) && ((nextSquares[10] === '🔴') || (nextSquares[10] === '❤️')) && (i === 1)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[5] === null) && ((nextSquares[12] === '🔴') || (nextSquares[12] === '❤️')) && (i === 5)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[26] === null) && (i === 26)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[28] === null) && (i === 28)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[33] === null) && ((nextSquares[26] === '🔴') || (nextSquares[26] === '❤️')) && (i === 33)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[37] === null) && ((nextSquares[28] === '🔴') || (nextSquares[28] === '❤️')) && (i === 37)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 19) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 21) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[12] === null) && (i === 12)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[14] === null) && (i === 14)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[3] === null) && ((nextSquares[12] === '🔴') || (nextSquares[12] === '❤️')) && (i === 3)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[7] === null) && ((nextSquares[14] === '🔴') || (nextSquares[14] === '❤️')) && (i === 7)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[28] === null) && (i === 28)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[30] === null) && (i === 30)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[35] === null) && ((nextSquares[28] === '🔴') || (nextSquares[28] === '❤️')) && (i === 35)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[39] === null) && ((nextSquares[30] === '🔴') || (nextSquares[30] === '❤️')) && (i === 39)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 21) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 23) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[14] === null) && (i === 14)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[5] === null) && ((nextSquares[14] === '🔴') || (nextSquares[14] === '❤️')) && (i === 5)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[30] === null) && (i === 30)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[37] === null) && ((nextSquares[30] === '🔴') || (nextSquares[30] === '❤️')) && (i === 37)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 23) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 24) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[17] === null) && (i === 17)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[10] === null) && ((nextSquares[17] === '🔴') || (nextSquares[17] === '❤️')) && (i === 10)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[33] === null) && (i === 33)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[42] === null) && ((nextSquares[33] === '🔴') || (nextSquares[33] === '❤️')) && (i === 42)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 24) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 26) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[17] === null) && (i === 17)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[19] === null) && (i === 19)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[8] === null) && ((nextSquares[17] === '🔴') || (nextSquares[17] === '❤️')) && (i === 8)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[12] === null) && ((nextSquares[19] === '🔴') || (nextSquares[19] === '❤️')) && (i === 12)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[33] === null) && (i === 33)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[35] === null) && (i === 35)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[40] === null) && ((nextSquares[33] === '🔴') || (nextSquares[33] === '❤️')) && (i === 40)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[44] === null) && ((nextSquares[35] === '🔴') || (nextSquares[35] === '❤️')) && (i === 44)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 26) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 28) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[19] === null) && (i === 19)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[21] === null) && (i === 21)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[10] === null) && ((nextSquares[19] === '🔴') || (nextSquares[19] === '❤️')) && (i === 10)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[14] === null) && ((nextSquares[21] === '🔴') || (nextSquares[21] === '❤️')) && (i === 14)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[35] === null) && (i === 35)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[37] === null) && (i === 37)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[42] === null) && ((nextSquares[35] === '🔴') || (nextSquares[35] === '❤️')) && (i === 42)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[46] === null) && ((nextSquares[37] === '🔴') || (nextSquares[37] === '❤️')) && (i === 46)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 28) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 30) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[21] === null) && (i === 21)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[23] === null) && (i === 23)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[12] === null) && ((nextSquares[21] === '🔴') || (nextSquares[21] === '❤️')) && (i === 12)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[37] === null) && (i === 37)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[39] === null) && (i === 39)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[44] === null) && ((nextSquares[37] === '🔴') || (nextSquares[37] === '❤️')) && (i === 44)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 30) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 33) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[24] === null) && (i === 24)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[26] === null) && (i === 26)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[19] === null) && ((nextSquares[26] === '🔴') || (nextSquares[26] === '❤️')) && (i === 19)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[40] === null) && (i === 40)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[42] === null) && (i === 42)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[51] === null) && ((nextSquares[42] === '🔴') || (nextSquares[42] === '❤️')) && (i === 51)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 33) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 35) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[26] === null) && (i === 26)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[28] === null) && (i === 28)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[17] === null) && ((nextSquares[26] === '🔴') || (nextSquares[26] === '❤️')) && (i === 17)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[21] === null) && ((nextSquares[28] === '🔴') || (nextSquares[28] === '❤️')) && (i === 21)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[42] === null) && (i === 42)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[44] === null) && (i === 44)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[49] === null) && ((nextSquares[42] === '🔴') || (nextSquares[42] === '❤️')) && (i === 49)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[53] === null) && ((nextSquares[44] === '🔴') || (nextSquares[44] === '❤️')) && (i === 53)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 35) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 37) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[28] === null) && (i === 28)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[30] === null) && (i === 30)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[19] === null) && ((nextSquares[28] === '🔴') || (nextSquares[28] === '❤️')) && (i === 19)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[23] === null) && ((nextSquares[30] === '🔴') || (nextSquares[30] === '❤️')) && (i === 23)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[44] === null) && (i === 44)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[46] === null) && (i === 46)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[51] === null) && ((nextSquares[44] === '🔴') || (nextSquares[44] === '❤️')) && (i === 51)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[55] === null) && ((nextSquares[46] === '🔴') || (nextSquares[46] === '❤️')) && (i === 55)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 37) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 39) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[30] === null) && (i === 30)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[21] === null) && ((nextSquares[30] === '🔴') || (nextSquares[30] === '❤️')) && (i === 21)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[46] === null) && (i === 46)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[53] === null) && ((nextSquares[46] === '🔴') || (nextSquares[46] === '❤️')) && (i === 53)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 39) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 40) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[33] === null) && (i === 33)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[26] === null) && ((nextSquares[33] === '🔴') || (nextSquares[33] === '❤️')) && (i === 26)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[49] === null) && (i === 49)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[58] === null) && ((nextSquares[49] === '🔴') || (nextSquares[49] === '❤️')) && (i === 58)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 40) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 42) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[33] === null) && (i === 33)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[35] === null) && (i === 35)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[24] === null) && ((nextSquares[33] === '🔴') || (nextSquares[33] === '❤️')) && (i === 24)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[28] === null) && ((nextSquares[35] === '🔴') || (nextSquares[35] === '❤️')) && (i === 28)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[49] === null) && (i === 49)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[51] === null) && (i === 51)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[56] === null) && ((nextSquares[49] === '🔴') || (nextSquares[49] === '❤️')) && (i === 56)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[60] === null) && ((nextSquares[51] === '🔴') || (nextSquares[51] === '❤️')) && (i === 60)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 42) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 44) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[35] === null) && (i === 35)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[37] === null) && (i === 37)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[26] === null) && ((nextSquares[35] === '🔴') || (nextSquares[35] === '❤️')) && (i === 26)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[30] === null) && ((nextSquares[37] === '🔴') || (nextSquares[37] === '❤️')) && (i === 30)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[51] === null) && (i === 51)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[53] === null) && (i === 53)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[58] === null) && ((nextSquares[51] === '🔴') || (nextSquares[51] === '❤️')) && (i === 58)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[62] === null) && ((nextSquares[53] === '🔴') || (nextSquares[53] === '❤️')) && (i === 62)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 44) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 46) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[37] === null) && (i === 37)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[39] === null) && (i === 39)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[28] === null) && ((nextSquares[37] === '🔴') || (nextSquares[37] === '❤️')) && (i === 28)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[53] === null) && (i === 53)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[55] === null) && (i === 55)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[60] === null) && ((nextSquares[53] === '🔴') || (nextSquares[53] === '❤️')) && (i === 60)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 46) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 49) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[40] === null) && (i === 40)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[42] === null) && (i === 42)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[35] === null) && ((nextSquares[42] === '🔴') || (nextSquares[42] === '❤️')) && (i === 35)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[56] === null) && (i === 56)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[58] === null) && (i === 58)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 49) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 51) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[42] === null) && (i === 42)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[44] === null) && (i === 44)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[33] === null) && ((nextSquares[42] === '🔴') || (nextSquares[42] === '❤️')) && (i === 33)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[37] === null) && ((nextSquares[44] === '🔴') || (nextSquares[44] === '❤️')) && (i === 37)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[58] === null) && (i === 58)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[60] === null) && (i === 60)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 51) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 53) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[44] === null) && (i === 44)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[46] === null) && (i === 46)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[35] === null) && ((nextSquares[44] === '🔴') || (nextSquares[44] === '❤️')) && (i === 35)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[39] === null) && ((nextSquares[46] === '🔴') || (nextSquares[46] === '❤️')) && (i === 39)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[60] === null) && (i === 60)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[62] === null) && (i === 62)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 53) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 55) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[46] === null) && (i === 46)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[37] === null) && ((nextSquares[46] === '🔴') || (nextSquares[46] === '❤️')) && (i === 37)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[62] === null) && (i === 62)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 55) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 56) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[49] === null) && (i === 49)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[42] === null) && ((nextSquares[49] === '🔴') || (nextSquares[49] === '❤️')) && (i === 42)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 56) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 58) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[49] === null) && (i === 49)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[51] === null) && (i === 51)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[40] === null) && ((nextSquares[49] === '🔴') || (nextSquares[49] === '❤️')) && (i === 40)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[44] === null) && ((nextSquares[51] === '🔴') || (nextSquares[51] === '❤️')) && (i === 44)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 58) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 60) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[51] === null) && (i === 51)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+          } else if ((nextSquares[53] === null) && (i === 53)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+          } else if ((nextSquares[42] === null) && ((nextSquares[51] === '🔴') || (nextSquares[51] === '❤️')) && (i === 42)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+          } else if ((nextSquares[46] === null) && ((nextSquares[53] === '🔴') || (nextSquares[53] === '❤️')) && (i === 46)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+          } else if (i === 60) {
+            nextSquares[i] = selectedPiece;
+          } else {
+            console.log('Invalid Move');
+          }
+      }
+      if ((ghostPosition === 62) && blackIsNext && (selectedPiece === '🖤')) {
+        if ((nextSquares[53] === null) && (i === 53)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+          } else if ((nextSquares[55] === null) && (i === 55)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+          } else if ((nextSquares[44] === null) && ((nextSquares[53] === '🔴') || (nextSquares[53] === '❤️')) && (i === 44)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+          } else if (i === 62) {
+            nextSquares[i] = selectedPiece;
+          } else {
+            console.log('Invalid Move');
+          }
+      }
+
+      // Single Red Piece Move Set
+      if ((ghostPosition === 8) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[1] === null) && (i === 1)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 8) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 10) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[1] === null) && (i === 1)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[3] === null) && (i === 3)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 10) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 12) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[3] === null) && (i === 3)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[5] === null) && (i === 5)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 12) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 14) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[5] === null) && (i === 5)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[7] === null) && (i === 7)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 14) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 17) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[8] === null) && (i === 8)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[10] === null) && (i === 10)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[3] === null) && ((nextSquares[10] === '⚫') || (nextSquares[10] === '🖤')) && (i === 3)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 17) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 19) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[10] === null) && (i === 10)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[12] === null) && (i === 12)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[1] === null) && ((nextSquares[10] === '⚫') || (nextSquares[10] === '🖤')) && (i === 1)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[5] === null) && ((nextSquares[12] === '⚫') || (nextSquares[12] === '🖤')) && (i === 5)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 19) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 21) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[12] === null) && (i === 12)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[14] === null) && (i === 14)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[3] === null) && ((nextSquares[12] === '⚫') || (nextSquares[12] === '🖤')) && (i === 3)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[7] === null) && ((nextSquares[14] === '⚫') || (nextSquares[14] === '🖤')) && (i === 7)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 21) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 23) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[14] === null) && (i === 14)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[5] === null) && ((nextSquares[14] === '⚫') || (nextSquares[14] === '🖤')) && (i === 5)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 23) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 24) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[17] === null) && (i === 17)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[10] === null) && ((nextSquares[17] === '⚫') || (nextSquares[17] === '🖤')) && (i === 10)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 24) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 26) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[17] === null) && (i === 17)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[19] === null) && (i === 19)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[8] === null) && ((nextSquares[17] === '⚫') || (nextSquares[17] === '🖤')) && (i === 8)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[12] === null) && ((nextSquares[19] === '⚫') || (nextSquares[19] === '🖤')) && (i === 12)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 26) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 28) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[19] === null) && (i === 19)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[21] === null) && (i === 21)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[10] === null) && ((nextSquares[19] === '⚫') || (nextSquares[19] === '🖤')) && (i === 10)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[14] === null) && ((nextSquares[21] === '⚫') || (nextSquares[21] === '🖤')) && (i === 14)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 28) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 30) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[21] === null) && (i === 21)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[23] === null) && (i === 23)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[12] === null) && ((nextSquares[21] === '⚫') || (nextSquares[21] === '🖤')) && (i === 12)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 30) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 33) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[24] === null) && (i === 24)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[26] === null) && (i === 26)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[19] === null) && ((nextSquares[26] === '⚫') || (nextSquares[26] === '🖤')) && (i === 19)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 33) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 35) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[26] === null) && (i === 26)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[28] === null) && (i === 28)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[17] === null) && ((nextSquares[26] === '⚫') || (nextSquares[26] === '🖤')) && (i === 17)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[21] === null) && ((nextSquares[28] === '⚫') || (nextSquares[28] === '🖤')) && (i === 21)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 35) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 37) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[28] === null) && (i === 28)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[30] === null) && (i === 30)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[19] === null) && ((nextSquares[28] === '⚫') || (nextSquares[28] === '🖤')) && (i === 19)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[23] === null) && ((nextSquares[30] === '⚫') || (nextSquares[30] === '🖤')) && (i === 23)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 37) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 39) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[30] === null) && (i === 30)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[21] === null) && ((nextSquares[30] === '⚫') || (nextSquares[30] === '🖤')) && (i === 21)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 39) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 40) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[33] === null) && (i === 33)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[26] === null) && ((nextSquares[33] === '⚫') || (nextSquares[33] === '🖤')) && (i === 26)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 40) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 42) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[33] === null) && (i === 33)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[35] === null) && (i === 35)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[24] === null) && ((nextSquares[33] === '⚫') || (nextSquares[33] === '🖤')) && (i === 24)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[28] === null) && ((nextSquares[35] === '⚫') || (nextSquares[35] === '🖤')) && (i === 28)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 42) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 44) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[35] === null) && (i === 35)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[37] === null) && (i === 37)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[26] === null) && ((nextSquares[35] === '⚫') || (nextSquares[35] === '🖤')) && (i === 26)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[30] === null) && ((nextSquares[37] === '⚫') || (nextSquares[37] === '🖤')) && (i === 30)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 44) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 46) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[37] === null) && (i === 37)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[39] === null) && (i === 39)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[28] === null) && ((nextSquares[37] === '⚫') || (nextSquares[37] === '🖤')) && (i === 28)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 46) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 49) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[40] === null) && (i === 40)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[42] === null) && (i === 42)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[35] === null) && ((nextSquares[42] === '⚫') || (nextSquares[42] === '🖤')) && (i === 35)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 49) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 51) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[42] === null) && (i === 42)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[44] === null) && (i === 44)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[33] === null) && ((nextSquares[42] === '⚫') || (nextSquares[42] === '🖤')) && (i === 33)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[37] === null) && ((nextSquares[44] === '⚫') || (nextSquares[44] === '🖤')) && (i === 37)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 51) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 53) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[44] === null) && (i === 44)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[46] === null) && (i === 46)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[35] === null) && ((nextSquares[44] === '⚫') || (nextSquares[44] === '🖤')) && (i === 35)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[39] === null) && ((nextSquares[46] === '⚫') || (nextSquares[46] === '🖤')) && (i === 39)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 53) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 55) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[46] === null) && (i === 46)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[37] === null) && ((nextSquares[46] === '⚫') || (nextSquares[46] === '🖤')) && (i === 37)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 55) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 56) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[49] === null) && (i === 49)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[42] === null) && ((nextSquares[49] === '⚫') || (nextSquares[49] === '🖤')) && (i === 42)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 56) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 58) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[49] === null) && (i === 49)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[51] === null) && (i === 51)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[40] === null) && ((nextSquares[49] === '⚫') || (nextSquares[49] === '🖤')) && (i === 40)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[44] === null) && ((nextSquares[51] === '⚫') || (nextSquares[51] === '🖤')) && (i === 44)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 58) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 60) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[51] === null) && (i === 51)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[53] === null) && (i === 53)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[42] === null) && ((nextSquares[51] === '⚫') || (nextSquares[51] === '🖤')) && (i === 42)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[46] === null) && ((nextSquares[53] === '⚫') || (nextSquares[53] === '🖤')) && (i === 46)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 60) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 62) && !blackIsNext && (selectedPiece === '🔴')) {
+        if ((nextSquares[53] === null) && (i === 53)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[55] === null) && (i === 55)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if ((nextSquares[44] === null) && ((nextSquares[53] === '⚫') || (nextSquares[53] === '🖤')) && (i === 44)) {
+          nextSquares[i] = selectedPiece;
+          removeGhostPieces();
+        } else if (i === 62) {
+          nextSquares[i] = selectedPiece;
+        } else {
+          console.log('Invalid Move');
+        }
+      }
+
+      // Kinged Red Piece Move Set
+      if ((ghostPosition === 1) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[8] === null) && (i === 8)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[10] === null) && (i === 10)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[19] === null) && ((nextSquares[10] === '⚫') || (nextSquares[10] === '🖤')) && (i === 19)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 1) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 3) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[10] === null) && (i === 10)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[12] === null) && (i === 12)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[17] === null) && ((nextSquares[10] === '⚫') || (nextSquares[10] === '🖤')) && (i === 17)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[21] === null) && ((nextSquares[12] === '⚫') || (nextSquares[12] === '🖤')) && (i === 21)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 3) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 5) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[12] === null) && (i === 12)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[14] === null) && (i === 14)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[19] === null) && ((nextSquares[12] === '⚫') || (nextSquares[12] === '🖤')) && (i === 19)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[23] === null) && ((nextSquares[14] === '⚫') || (nextSquares[14] === '🖤')) && (i === 23)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 5) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 7) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[14] === null) && (i === 14)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[21] === null) && ((nextSquares[14] === '⚫') || (nextSquares[14] === '🖤')) && (i === 21)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 7) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 8) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[1] === null) && (i === 1)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[17] === null) && (i === 17)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[26] === null) && ((nextSquares[17] === '⚫') || (nextSquares[17] === '🖤')) && (i === 26)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 8) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 10) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[1] === null) && (i === 1)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[3] === null) && (i === 3)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[17] === null) && (i === 17)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[19] === null) && (i === 19)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[24] === null) && ((nextSquares[17] === '⚫') || (nextSquares[17] === '🖤')) && (i === 24)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[28] === null) && ((nextSquares[19] === '⚫') || (nextSquares[19] === '🖤')) && (i === 28)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 10) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 12) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[3] === null) && (i === 3)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[5] === null) && (i === 5)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[19] === null) && (i === 19)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[21] === null) && (i === 21)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[26] === null) && ((nextSquares[19] === '⚫') || (nextSquares[19] === '🖤')) && (i === 26)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[30] === null) && ((nextSquares[21] === '⚫') || (nextSquares[21] === '🖤')) && (i === 30)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 12) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 14) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[5] === null) && (i === 5)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[7] === null) && (i === 7)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[21] === null) && (i === 21)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[23] === null) && (i === 23)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[28] === null) && ((nextSquares[21] === '⚫') || (nextSquares[21] === '🖤')) && (i === 28)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 14) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 17) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[8] === null) && (i === 8)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[10] === null) && (i === 10)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[3] === null) && ((nextSquares[10] === '⚫') || (nextSquares[10] === '🖤')) && (i === 3)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[24] === null) && (i === 24)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[26] === null) && (i === 26)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[35] === null) && ((nextSquares[26] === '⚫') || (nextSquares[26] === '🖤')) && (i === 35)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 17) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 19) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[10] === null) && (i === 10)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[12] === null) && (i === 12)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[1] === null) && ((nextSquares[10] === '⚫') || (nextSquares[10] === '🖤')) && (i === 1)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[5] === null) && ((nextSquares[12] === '⚫') || (nextSquares[12] === '🖤')) && (i === 5)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[26] === null) && (i === 26)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[28] === null) && (i === 28)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[33] === null) && ((nextSquares[26] === '⚫') || (nextSquares[26] === '🖤')) && (i === 33)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[37] === null) && ((nextSquares[28] === '⚫') || (nextSquares[28] === '🖤')) && (i === 37)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 19) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 21) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[12] === null) && (i === 12)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[14] === null) && (i === 14)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[3] === null) && ((nextSquares[12] === '⚫') || (nextSquares[12] === '🖤')) && (i === 3)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[7] === null) && ((nextSquares[14] === '⚫') || (nextSquares[14] === '🖤')) && (i === 7)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[28] === null) && (i === 28)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[30] === null) && (i === 30)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[35] === null) && ((nextSquares[28] === '⚫') || (nextSquares[28] === '🖤')) && (i === 35)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[39] === null) && ((nextSquares[30] === '⚫') || (nextSquares[30] === '🖤')) && (i === 39)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 21) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 23) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[14] === null) && (i === 14)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[5] === null) && ((nextSquares[14] === '⚫') || (nextSquares[14] === '🖤')) && (i === 5)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[30] === null) && (i === 30)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[37] === null) && ((nextSquares[30] === '⚫') || (nextSquares[30] === '🖤')) && (i === 37)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 23) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 24) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[17] === null) && (i === 17)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[10] === null) && ((nextSquares[17] === '⚫') || (nextSquares[17] === '🖤')) && (i === 10)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[33] === null) && (i === 33)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[42] === null) && ((nextSquares[33] === '⚫') || (nextSquares[33] === '🖤')) && (i === 42)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 24) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 26) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[17] === null) && (i === 17)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[19] === null) && (i === 19)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[8] === null) && ((nextSquares[17] === '⚫') || (nextSquares[17] === '🖤')) && (i === 8)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[12] === null) && ((nextSquares[19] === '⚫') || (nextSquares[19] === '🖤')) && (i === 12)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[33] === null) && (i === 33)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[35] === null) && (i === 35)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[40] === null) && ((nextSquares[33] === '⚫') || (nextSquares[33] === '🖤')) && (i === 40)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[44] === null) && ((nextSquares[35] === '⚫') || (nextSquares[35] === '🖤')) && (i === 44)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 26) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 28) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[19] === null) && (i === 19)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[21] === null) && (i === 21)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[10] === null) && ((nextSquares[19] === '⚫') || (nextSquares[19] === '🖤')) && (i === 10)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[14] === null) && ((nextSquares[21] === '⚫') || (nextSquares[21] === '🖤')) && (i === 14)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[35] === null) && (i === 35)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[37] === null) && (i === 37)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[42] === null) && ((nextSquares[35] === '⚫') || (nextSquares[35] === '🖤')) && (i === 42)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[46] === null) && ((nextSquares[37] === '⚫') || (nextSquares[37] === '🖤')) && (i === 46)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 28) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 30) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[21] === null) && (i === 21)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[23] === null) && (i === 23)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[12] === null) && ((nextSquares[21] === '⚫') || (nextSquares[21] === '🖤')) && (i === 12)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[37] === null) && (i === 37)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[39] === null) && (i === 39)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[44] === null) && ((nextSquares[37] === '⚫') || (nextSquares[37] === '🖤')) && (i === 44)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 30) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 33) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[24] === null) && (i === 24)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[26] === null) && (i === 26)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[19] === null) && ((nextSquares[26] === '⚫') || (nextSquares[26] === '🖤')) && (i === 19)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[40] === null) && (i === 40)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[42] === null) && (i === 42)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[51] === null) && ((nextSquares[42] === '⚫') || (nextSquares[42] === '🖤')) && (i === 51)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 33) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 35) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[26] === null) && (i === 26)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[28] === null) && (i === 28)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[17] === null) && ((nextSquares[26] === '⚫') || (nextSquares[26] === '🖤')) && (i === 17)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[21] === null) && ((nextSquares[28] === '⚫') || (nextSquares[28] === '🖤')) && (i === 21)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[42] === null) && (i === 42)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[44] === null) && (i === 44)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[49] === null) && ((nextSquares[42] === '⚫') || (nextSquares[42] === '🖤')) && (i === 49)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[53] === null) && ((nextSquares[44] === '⚫') || (nextSquares[44] === '🖤')) && (i === 53)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 35) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 37) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[28] === null) && (i === 28)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[30] === null) && (i === 30)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[19] === null) && ((nextSquares[28] === '⚫') || (nextSquares[28] === '🖤')) && (i === 19)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[23] === null) && ((nextSquares[30] === '⚫') || (nextSquares[30] === '🖤')) && (i === 23)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[44] === null) && (i === 44)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[46] === null) && (i === 46)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[51] === null) && ((nextSquares[44] === '⚫') || (nextSquares[44] === '🖤')) && (i === 51)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[55] === null) && ((nextSquares[46] === '⚫') || (nextSquares[46] === '🖤')) && (i === 55)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 37) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 39) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[30] === null) && (i === 30)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[21] === null) && ((nextSquares[30] === '⚫') || (nextSquares[30] === '🖤')) && (i === 21)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[46] === null) && (i === 46)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[53] === null) && ((nextSquares[46] === '⚫') || (nextSquares[46] === '🖤')) && (i === 53)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 39) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 40) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[33] === null) && (i === 33)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[26] === null) && ((nextSquares[33] === '⚫') || (nextSquares[33] === '🖤')) && (i === 26)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[49] === null) && (i === 49)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[58] === null) && ((nextSquares[49] === '⚫') || (nextSquares[49] === '🖤')) && (i === 58)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 40) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 42) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[33] === null) && (i === 33)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[35] === null) && (i === 35)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[24] === null) && ((nextSquares[33] === '⚫') || (nextSquares[33] === '🖤')) && (i === 24)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[28] === null) && ((nextSquares[35] === '⚫') || (nextSquares[35] === '🖤')) && (i === 28)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[49] === null) && (i === 49)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[51] === null) && (i === 51)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[56] === null) && ((nextSquares[49] === '⚫') || (nextSquares[49] === '🖤')) && (i === 56)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[60] === null) && ((nextSquares[51] === '⚫') || (nextSquares[51] === '🖤')) && (i === 60)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 42) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 44) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[35] === null) && (i === 35)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[37] === null) && (i === 37)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[26] === null) && ((nextSquares[35] === '⚫') || (nextSquares[35] === '🖤')) && (i === 26)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[30] === null) && ((nextSquares[37] === '⚫') || (nextSquares[37] === '🖤')) && (i === 30)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[51] === null) && (i === 51)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[53] === null) && (i === 53)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[58] === null) && ((nextSquares[51] === '⚫') || (nextSquares[51] === '🖤')) && (i === 58)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[62] === null) && ((nextSquares[53] === '⚫') || (nextSquares[53] === '🖤')) && (i === 62)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 44) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 46) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[37] === null) && (i === 37)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[39] === null) && (i === 39)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[28] === null) && ((nextSquares[37] === '⚫') || (nextSquares[37] === '🖤')) && (i === 28)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[53] === null) && (i === 53)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[55] === null) && (i === 55)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[60] === null) && ((nextSquares[53] === '⚫') || (nextSquares[53] === '🖤')) && (i === 60)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 46) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 49) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[40] === null) && (i === 40)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[42] === null) && (i === 42)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[35] === null) && ((nextSquares[42] === '⚫') || (nextSquares[42] === '🖤')) && (i === 35)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[56] === null) && (i === 56)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[58] === null) && (i === 58)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 49) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 51) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[42] === null) && (i === 42)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[44] === null) && (i === 44)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[33] === null) && ((nextSquares[42] === '⚫') || (nextSquares[42] === '🖤')) && (i === 33)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[37] === null) && ((nextSquares[44] === '⚫') || (nextSquares[44] === '🖤')) && (i === 37)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[58] === null) && (i === 58)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[60] === null) && (i === 60)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 51) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 53) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[44] === null) && (i === 44)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[46] === null) && (i === 46)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[35] === null) && ((nextSquares[44] === '⚫') || (nextSquares[44] === '🖤')) && (i === 35)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[39] === null) && ((nextSquares[46] === '⚫') || (nextSquares[46] === '🖤')) && (i === 39)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[60] === null) && (i === 60)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[62] === null) && (i === 62)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 53) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 55) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[46] === null) && (i === 46)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[37] === null) && ((nextSquares[46] === '⚫') || (nextSquares[46] === '🖤')) && (i === 37)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[62] === null) && (i === 62)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 55) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 56) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[49] === null) && (i === 49)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[42] === null) && ((nextSquares[49] === '⚫') || (nextSquares[49] === '🖤')) && (i === 42)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 56) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 58) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[49] === null) && (i === 49)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[51] === null) && (i === 51)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[40] === null) && ((nextSquares[49] === '⚫') || (nextSquares[49] === '🖤')) && (i === 40)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if ((nextSquares[44] === null) && ((nextSquares[51] === '⚫') || (nextSquares[51] === '🖤')) && (i === 44)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+        } else if (i === 58) {
+            nextSquares[i] = selectedPiece;
+        } else {
+            console.log('Invalid Move');
+        }
+      }
+      if ((ghostPosition === 60) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[51] === null) && (i === 51)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+          } else if ((nextSquares[53] === null) && (i === 53)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+          } else if ((nextSquares[42] === null) && ((nextSquares[51] === '⚫') || (nextSquares[51] === '🖤')) && (i === 42)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+          } else if ((nextSquares[46] === null) && ((nextSquares[53] === '⚫') || (nextSquares[53] === '🖤')) && (i === 46)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+          } else if (i === 60) {
+            nextSquares[i] = selectedPiece;
+          } else {
+            console.log('Invalid Move');
+          }
+      }
+      if ((ghostPosition === 62) && !blackIsNext && (selectedPiece === '❤️')) {
+        if ((nextSquares[53] === null) && (i === 53)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+          } else if ((nextSquares[55] === null) && (i === 55)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+          } else if ((nextSquares[44] === null) && ((nextSquares[53] === '⚫') || (nextSquares[53] === '🖤')) && (i === 44)) {
+            nextSquares[i] = selectedPiece;
+            removeGhostPieces();
+          } else if (i === 62) {
+            nextSquares[i] = selectedPiece;
+          } else {
+            console.log('Invalid Move');
+          }
+      }
+    }
   
     if ((nextSquares[i] === '🔴') && !blackIsNext) {
       changeSelectedPiece();
-      console.log(selectedPiece);
       nextSquares[i] = 'R';
+      getGhostPosition();
       tallyPieces();
     } else if ((nextSquares[i] === '❤️') && !blackIsNext) {
       changeSelectedPiece();
-      console.log(selectedPiece);
       nextSquares[i] = 'RK';
+      getGhostPosition();
       tallyPieces();
     } else if ((nextSquares[i] === '⚫') && blackIsNext) {
       changeSelectedPiece();
-      console.log(selectedPiece);
       nextSquares[i] = 'B';
+      getGhostPosition();
       tallyPieces();
     } else if ((nextSquares[i] === '🖤') && blackIsNext) {
       changeSelectedPiece();
-      console.log(selectedPiece);
       nextSquares[i] = 'BK';
+      getGhostPosition();
       tallyPieces();
     } else if ((nextSquares[i] === '🔴') || (nextSquares[i] === '❤️') || (nextSquares[i] === '⚫') || (nextSquares[i] === '🖤')) {
       changeSelectedPiece();
-      console.log(selectedPiece);
       nextSquares[i] = null;
       tallyPieces();
     } else if (((nextSquares[i] === null) || (nextSquares[i] === 'B') || (nextSquares[i] === 'BK'))  && blackIsNext) {
-      nextSquares[i] = selectedPiece;
-      removeGhostPieces();
+      movementRules();
       if (nextSquares[56] === '⚫') {
         (nextSquares[56] = '🖤')
-        // Add 1 to black's kinged piece stat
+        // TODO Add 1 to black's kinged piece stat
       }
       if (nextSquares[58] === '⚫') {
         (nextSquares[58] = '🖤')
-        // Add 1 to black's kinged piece stat
+        // TODO Add 1 to black's kinged piece stat
       }
       if (nextSquares[60] === '⚫') {
         (nextSquares[60] = '🖤')
-        // Add 1 to black's kinged piece stat
+        // TODO Add 1 to black's kinged piece stat
       }
       if (nextSquares[62] === '⚫') {
         (nextSquares[62] = '🖤')
-        // Add 1 to black's kinged piece stat
+        // TODO Add 1 to black's kinged piece stat
       }
       tallyPieces();
     } else if (((nextSquares[i] === null) || (nextSquares[i] === 'R') || (nextSquares[i] === 'RK')) && !blackIsNext) {
-      nextSquares[i] = selectedPiece;
-      removeGhostPieces();
+      movementRules();
       if (nextSquares[1] === '🔴') {
         (nextSquares[1] = '❤️')
-        // Add 1 to red's kinged piece stat
+        // TODO Add 1 to red's kinged piece stat
       }
       if (nextSquares[3] === '🔴') {
         (nextSquares[3] = '❤️')
-        // Add 1 to red's kinged piece stat
+        // TODO Add 1 to red's kinged piece stat
       }
       if (nextSquares[5] === '🔴') {
         (nextSquares[5] = '❤️')
-        // Add 1 to red's kinged piece stat
+        // TODO Add 1 to red's kinged piece stat
       }
       if (nextSquares[7] === '🔴') {
         (nextSquares[7] = '❤️')
-        // Add 1 to red's kinged piece stat
+        // TODO Add 1 to red's kinged piece stat
       }      
       tallyPieces();
     }
@@ -247,8 +2639,6 @@ export default function Board() {
     </>
   );
 }
-
-
